server/fundoo_notes/requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
loguru==0.7.2
django-redis==5.4.0
celery==5.4.0
flower==2.0.1
django-celery-beat==2.7.0
drf-yasg==1.21.7
pytest-django==4.9.0
pytest==8.3.2
pytest-cov==5.0.0